♻️ Move initial actions to their own file.
